DESENG-801: Move Widget code out of old view folder
