+++ met-web/src/components/engagement/widgets/Events/InPersonEvent.tsx
